qa/crontab: update frequency and priority for rados nightlies
Although main rados nightlies are scheduled to run weekly, they
are so low in the queue, that they take weeks to run, and when
they finally do, all the jobs die (likely due to waiting in the
queue so long). This change ups the priority for main runs, since
this is most important for debugging new failures. To compliment this
increase in priority, I also decreased the priority of squid nightly
rados runs from twice a week to once a week since that is more than
enough.

Signed-off-by: Laura Flores <lflores@ibm.com>