osd: break relevant code into a new `try_upmap` function
The whole section that `try_upmap` encompasses loops over
all pgs sequentially to perform operations on them. The motivation
behind moving this section to its own function is to make the code
more readable, and also to set up the code for multithreading, which
will need to be spread among all the pgs in multiple threads instead
of one sequential for loop.

Signed-off-by: Laura Flores <lflores@ibm.com>