osd: add pg-upmap-primary to clean_pg_upmaps
clean_pg_upmaps is used to check for invalid upmaps. For
instance, if pool size is decreased and a pg with a mapping
disappears, the mapping is invalid.

This commit adds handling of pg_upmap_primary to clean_pg_upmaps
so mappings are properly removed if a pool size is decreased.

Fixes: https://tracker.ceph.com/issues/67265
Signed-off-by: Laura Flores <lflores@ibm.com>