mgr/balancer: Warn about pg-upmap-primary mappings when switching modes
Currently, if a user enables a balancer module mode that uses
"pg-upmap-primary" (`upmap-read` or `read`), but then decides to switch
back to a mode that does not use "pg-upmap-primary" ("upmap" or
"crush-compat"), no further read balancing occurs, but the pg-upmap-primary
mappings remain in the osdmap.

With this patch, the balancer module is enhanced to alert users about
pg-upmap-primary mappings in their cluster upon switching to "upmap" or
"crush-compat" mode. This is especially helpful when a user might want to
use a client older than Reef their cluster, which will not be allowed if
pg-upmap-primary mappings are still in use. Any attempt to set
require_min_compat_client below reef (i.e. luminous) will result in the
following error:

```
Error EPERM: osdmap current utilizes features that require reef; cannot
set require_min_compat_client below that to luminous
```

Now the user is alerted in case they want to take action to remove the
mappings:

```
$ ceph balancer mode upmap
NOTICE: You have switched to a mode that does not use pg-upmap-primary,
but your osdmap still contains pg-upmap-primary mappings.

If you want to remove these mappings, run `ceph osd rm-pg-upmap-primary-all`.
Otherwise, the mappings will be retained, but read balancing will no longer
take place.
```

Instead of silently removing the mappings in the background, this
notice allows for any case in which a user may want to keep the mappings
when switching to another mode.

Fixes: https://tracker.ceph.com/issues/70615
Signed-off-by: Laura Flores <lflores@ibm.com>