qa/suites: enable 'watchdog_setup' for thrash-old-clients and upgrade tests
Opting in to 'watchdog_setup' would be beneficial for the thrash-old-clients
and upgrade tests since the default watchdog timeout of 300 seconds should prevent
the watchdog from timing out as cephadm stops and restarts daemons. Also, we want
the protection of the watchdog to stop jobs where OSDs assert rathter than hanging
the job until it is marked dead after 8 hours.

Signed-off-by: Laura Flores <lflores@ibm.com>